initial implementation of align algorithm in PGrid

+++ include/Range.hpp
@@ -6,15 +6,15 @@ namespace dvmpredictor {
 	class Range {
 	public:
 		Range();
-		Range(int32_t start, uint32_t count, bool forward = true);
+		Range(int64_t start, uint64_t count, bool forward = true);
 
-		int32_t start() const;
-		uint32_t count() const;
+		int64_t start() const;
+		uint64_t count() const;
 		bool forward() const;
 
 	private:
-		int32_t _start;
-		uint32_t _count;
+		int64_t _start;
+		uint64_t _count;
 		bool _forward;
 	};
 }
